requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
aiofiles
requests
Pillow
nbstripout==0.8.1
PyAudio==0.2.14
SpeechRecognition==3.14.3
spacy>=3.0.0
transformers>=4.0.0
google-generativeai
torch
